:recycle: refactor: MediaCard agora é compartilhado
