+++ app/src/componentes/MediaCard/MediaCard.tsx
